requirements.txt
fastapi>=0.115
uvicorn>=0.34
ollama>=0.4
psycopg2-binary>=2.9
pygame>=2.5
SpeechRecognition>=3.10
edge-tts>=6.0
duckduckgo-search>=6.0
keyboard>=0.13
python-dotenv>=1.0
yfinance>=0.2
google-auth-oauthlib>=1.0
google-api-python-client>=2.0
psutil>=5.9
requests>=2.31
PyAudio>=0.2
Pillow>=10.0
pydantic>=2.0
faster-whisper>=1.0
RestrictedPython>=7.0
